Added joint_position and joint_message classes to prepare for sending joint trajectories between ROS and the robot controller.

+++ trunk/industrial/simple_message/include/simple_serialize.h
@@ -77,7 +77,7 @@ public:
    *
    * \return object size (in bytes)
    */
-  virtual size_t byteLength()=0;
+  virtual unsigned int byteLength()=0;
 
 };
